requirements.txt
matplotlib>=3.5
numpy>=1.21
requests>=2.28
python-dotenv>=1.0
